refactor send e-mail
