--- scripts/yo.html
@@ -1,148 +1,0 @@
-<h1 id="near-merged-pull-requests-for-october-2023">NEAR Merged Pull Requests for October 2023</h1><h2 id="table-of-contents">Table of Contents</h2><ul>
-    <li><a href="#docs">DOCS</a> </li>
-    <li><a href="#near-discovery">NEAR-DISCOVERY</a></li>
-    </ul>
-    <hr>
-    <h2 id="docs">DOCS</h2><table border="0" cellpadding="4" cellspacing="5">
-    <thead>
-    <tr>
-    <th>merged_at</th>
-    <th>num</th>
-    <th>title</th>
-    </tr>
-    </thead>
-    <tbody><tr>
-    <td>2023-10-29</td>
-    <td><a href="https://github.com/near/docs/pull/1558">1558</a></td>
-    <td><a href="https://github.com/near/docs/pull/1558">Typo fix in nfts.md</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-29</td>
-    <td><a href="https://github.com/near/docs/pull/1559">1559</a></td>
-    <td><a href="https://github.com/near/docs/pull/1559">Typo fix in anatomy.md</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-27</td>
-    <td><a href="https://github.com/near/docs/pull/1554">1554</a></td>
-    <td><a href="https://github.com/near/docs/pull/1554">Best Practices for Ethereum developers on BOS...</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-27</td>
-    <td><a href="https://github.com/near/docs/pull/1556">1556</a></td>
-    <td><a href="https://github.com/near/docs/pull/1556">Fixed Lido component link (BOS for Ethereum develo...</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-23</td>
-    <td><a href="https://github.com/near/docs/pull/1546">1546</a></td>
-    <td><a href="https://github.com/near/docs/pull/1546">Add a tutorial for websockets based on the widget ...</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-19</td>
-    <td><a href="https://github.com/near/docs/pull/1544">1544</a></td>
-    <td><a href="https://github.com/near/docs/pull/1544">feat: BOS for Ethereum developers doc</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-18</td>
-    <td><a href="https://github.com/near/docs/pull/1548">1548</a></td>
-    <td><a href="https://github.com/near/docs/pull/1548">Bump @babel/traverse from 7.18.13 to 7.23.2 in /we...</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-18</td>
-    <td><a href="https://github.com/near/docs/pull/1527">1527</a></td>
-    <td><a href="https://github.com/near/docs/pull/1527">Bump postcss from 8.4.16 to 8.4.31 in /website...</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-13</td>
-    <td><a href="https://github.com/near/docs/pull/1541">1541</a></td>
-    <td><a href="https://github.com/near/docs/pull/1541">update epic template</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-13</td>
-    <td><a href="https://github.com/near/docs/pull/1540">1540</a></td>
-    <td><a href="https://github.com/near/docs/pull/1540">add epic template</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-12</td>
-    <td><a href="https://github.com/near/docs/pull/1536">1536</a></td>
-    <td><a href="https://github.com/near/docs/pull/1536">update queryApi links</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-11</td>
-    <td><a href="https://github.com/near/docs/pull/1535">1535</a></td>
-    <td><a href="https://github.com/near/docs/pull/1535">Update simple-fts.md</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-10</td>
-    <td><a href="https://github.com/near/docs/pull/1526">1526</a></td>
-    <td><a href="https://github.com/near/docs/pull/1526">Update context.md - add log</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-06</td>
-    <td><a href="https://github.com/near/docs/pull/1525">1525</a></td>
-    <td><a href="https://github.com/near/docs/pull/1525">Remove FetchFromSocialApi </a></td>
-    </tr>
-    <tr>
-    <td>2023-10-05</td>
-    <td><a href="https://github.com/near/docs/pull/1522">1522</a></td>
-    <td><a href="https://github.com/near/docs/pull/1522">Update relayers doc</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-05</td>
-    <td><a href="https://github.com/near/docs/pull/1524">1524</a></td>
-    <td><a href="https://github.com/near/docs/pull/1524">Fix typos</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-04</td>
-    <td><a href="https://github.com/near/docs/pull/1523">1523</a></td>
-    <td><a href="https://github.com/near/docs/pull/1523">Gateway notifications</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-02</td>
-    <td><a href="https://github.com/near/docs/pull/1510">1510</a></td>
-    <td><a href="https://github.com/near/docs/pull/1510">update QueryApi limitations</a></td>
-    </tr>
-    </tbody></table>
-    <h2 id="near-discovery">NEAR-DISCOVERY</h2><table border="0" cellpadding="4" cellspacing="5">
-    <thead>
-    <tr>
-    <th>merged_at</th>
-    <th>num</th>
-    <th>title</th>
-    </tr>
-    </thead>
-    <tbody><tr>
-    <td>2023-10-26</td>
-    <td><a href="https://github.com/near/near-discovery/pull/769">769</a></td>
-    <td><a href="https://github.com/near/near-discovery/pull/769">weekly promotion of develop to main</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-24</td>
-    <td><a href="https://github.com/near/near-discovery/pull/761">761</a></td>
-    <td><a href="https://github.com/near/near-discovery/pull/761">develop -&gt; main</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-19</td>
-    <td><a href="https://github.com/near/near-discovery/pull/734">734</a></td>
-    <td><a href="https://github.com/near/near-discovery/pull/734">weekly promotion of develop to main</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-16</td>
-    <td><a href="https://github.com/near/near-discovery/pull/721">721</a></td>
-    <td><a href="https://github.com/near/near-discovery/pull/721">Upgrade vm  (#720)</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-16</td>
-    <td><a href="https://github.com/near/near-discovery/pull/719">719</a></td>
-    <td><a href="https://github.com/near/near-discovery/pull/719">develop -&gt; main</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-12</td>
-    <td><a href="https://github.com/near/near-discovery/pull/694">694</a></td>
-    <td><a href="https://github.com/near/near-discovery/pull/694">weekly promotion of develop to main</a></td>
-    </tr>
-    <tr>
-    <td>2023-10-04</td>
-    <td><a href="https://github.com/near/near-discovery/pull/654">654</a></td>
-    <td><a href="https://github.com/near/near-discovery/pull/654">develop -&gt; main</a></td>
-    </tr>
-    </tbody></table>
